Optimize program size

+++ FStopTestMode.h
@@ -1,6 +1,6 @@
 #include "ModeProcessor.h"
 
-#include <stddef.h>
+#include <stdint.h>
 
 class FStopTestMode final : public ModeProcessor {
     enum class Step { initTime, fstopSet, run, log, last_ };
@@ -12,5 +12,5 @@ public:
 
 private:
     Step m_step;
-    size_t m_currentRun;
+    uint8_t m_currentRun;
 };
