+++ MaskMode.h
@@ -1,6 +1,6 @@
 #include "ModeProcessor.h"
 
-#include <stddef.h>
+#include <stdint.h>
 
 class MaskMode final : public ModeProcessor {
     enum class Step { setNum, setMasks, log, run, last_ };
@@ -12,6 +12,6 @@ public:
 
 private:
     Step m_step;
-    size_t m_numberOfMasks;
-    size_t m_currentMask;
+    uint8_t m_numberOfMasks;
+    uint8_t m_currentMask;
 };
